requirements.txt
typing==3.7.4.3
python-dotenv==1.0.0
pandas==2.1.1
openai==0.28.1
langchain==0.0.306
google-api-core==2.12.0
google-cloud-documentai==2.20.0